Add ability to use eco templates
Purposely did not add the ruby-eco gem dependency
to the gemspec because I figured it would be best
to only rely on it if, in assets.yml, you do:
template_function: off
template_extension: eco

This is a rough whack at getting it to work, I didn't
add any tests or anything.  Is this something
you want to do?